requirements.txt
fastapi==0.78.0
strawberry-graphql==0.17.0
uvicorn==0.18.2
